+++ python/datafusion/__init__.py
@@ -49,7 +49,15 @@ from .html_formatter import configure_formatter
 from .io import read_avro, read_csv, read_json, read_parquet
 from .plan import ExecutionPlan, LogicalPlan
 from .record_batch import RecordBatch, RecordBatchStream
-from .udf import Accumulator, AggregateUDF, ScalarUDF, WindowUDF, udaf, udf, udwf
+from .user_defined import (
+    Accumulator,
+    AggregateUDF,
+    ScalarUDF,
+    WindowUDF,
+    udaf,
+    udf,
+    udwf,
+)
 
 __version__ = importlib_metadata.version(__name__)
