+++ python/tests/test_udwf.py
@@ -22,7 +22,7 @@ import pytest
 from datafusion import SessionContext, column, lit, udwf
 from datafusion import functions as f
 from datafusion.expr import WindowFrame
-from datafusion.udf import WindowEvaluator
+from datafusion.user_defined import WindowEvaluator
 
 
 class ExponentialSmoothDefault(WindowEvaluator):
